refactor(api-e2e): split entity-helpers fixture into per-entity modules
Bean: api-k307
